requirements.txt
streamlit>=1.24.0
pandas>=1.5.3
numpy>=1.24.3
plotly>=5.14.1
python-dotenv>=1.0.0
requests>=2.31.0
openpyxl>=3.1.2
xlsxwriter>=3.1.0
python-dateutil>=2.8.2
pytz>=2023.3
tzdata>=2023.3
